backend/requirements.txt
django
djangorestframework
psycopg
python-decouple
django-cors-headers
drf-nested-routers